REF: Refactor for TypeScript
